🧹 move shared apps functions to the root file
