+++ apps/cnquery/cmd/run.go
@@ -2,7 +2,6 @@ package cmd
 
 import (
 	"os"
-	"time"
 
 	"github.com/cockroachdb/errors"
 	"github.com/rs/zerolog/log"
@@ -10,10 +9,6 @@ import (
 	"github.com/spf13/viper"
 	"go.mondoo.com/cnquery"
 	"go.mondoo.com/cnquery/apps/cnquery/cmd/builder"
-	"go.mondoo.com/cnquery/cli/inventoryloader"
-	"go.mondoo.com/cnquery/motor"
-	"go.mondoo.com/cnquery/motor/asset"
-	v1 "go.mondoo.com/cnquery/motor/inventory/v1"
 	"go.mondoo.com/cnquery/motor/providers"
 )
 
@@ -134,60 +129,4 @@ func GetCobraRunConfig(cmd *cobra.Command, args []string, provider providers.Pro
 	conf.PlatformID = viper.GetString("platform-id")
 
 	return &conf, nil
-}
-
-// TODO: consider moving this to inventoryloader package
-func getInventory(cliAsset *asset.Asset, insecure bool) (*v1.Inventory, error) {
-	var v1inventory *v1.Inventory
-	var err error
-
-	// parses optional inventory file if inventory was not piped already
-	if v1inventory == nil {
-		v1inventory, err = inventoryloader.Parse()
-		if err != nil {
-			return nil, errors.Wrap(err, "could not parse inventory")
-		}
-	}
-
-	// add asset from cli to inventory
-	if (len(v1inventory.Spec.GetAssets()) == 0) && cliAsset != nil {
-		v1inventory.AddAssets(cliAsset)
-	}
-
-	// if the --insecure flag is set, we overwrite the individual setting for the asset
-	if insecure == true {
-		v1inventory.MarkConnectionsInsecure()
-	}
-
-	return v1inventory, nil
-}
-
-func filterAssetByPlatformID(assetList []*asset.Asset, selectionID string) (*asset.Asset, error) {
-	var foundAsset *asset.Asset
-	for i := range assetList {
-		assetObj := assetList[i]
-		for j := range assetObj.PlatformIds {
-			if assetObj.PlatformIds[j] == selectionID {
-				return assetObj, nil
-			}
-		}
-	}
-
-	if foundAsset == nil {
-		return nil, errors.New("could not find an asset with the provided identifer: " + selectionID)
-	}
-	return foundAsset, nil
-}
-
-// storeRecording stores tracked commands and files into the recording file
-func storeRecording(m *motor.Motor) {
-	if m.IsRecording() {
-		filename := viper.GetString("record-file")
-		if filename == "" {
-			filename = "recording-" + time.Now().Format("20060102150405") + ".toml"
-		}
-		log.Info().Str("filename", filename).Msg("store recordings")
-		data := m.Recording()
-		os.WriteFile(filename, data, 0o700)
-	}
 }
